% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMII_functions.R
\name{CMI}
\alias{CMI}
\title{Conditional Mutual Information}
\usage{
CMI(X,Y,Z,method=c("covariance"),unit=c("bits",
"nats","hartley","normalized"),pvalue=FALSE,permutation_times=100)
}
\arguments{
\item{X}{a numeric vector to test}

\item{Y}{a numeric vector to test}

\item{Z}{a numeric vector as the condition}

\item{method}{the estimator method to test the CMI: "covariance"}

\item{unit}{The unit of the result: "bits", "nats", "hartley" and "normalized" (the default is "bits"). The normalized result will be between 0 and 1.}

\item{pvalue}{a logical value to determine whether to calculate the pvalue or not}

\item{permutation_times}{integral value to determin the permutation times in calculating p value.}
}
\value{
a numeric value of conditional mutual information between X and Y based on condition of Z
}
\description{
\code{CMI} takes three continuous variables as input and calculate the conditional mutual information between X and Y based on the condition of Z. Different from estimator method based on data discretization, this fucntion will use covarians transformation to estimate the continuous probabilities distribution of x and y values.
}
\examples{
x<-rnorm(100)
y<-0.7*x+rnorm(100,sd=0.1)
z<-0.8*x+rnorm(100,sd=0.1)
cor(x,y);cor(x,y);cor(y,z)  #correlation test cannot identify the direct connection
CMI(x,y,z)  # CMI identify the direct connection between x and y is not relying on
            #  the condiction of z
CMI(y,z,x)  # CMI identify the direct connection between y and z is not relying on
            #  the condiction of x
CMI(x,z,y)  # CMI identify the connection between x and z is depending on the condition of y
CMI(x,y,z,pvalue=TRUE)$adj.pvalue
CMI(x,z,y,pvalue=TRUE)$adj.pvalue
}
\references{
Zhang, X. (2011). Inferring gene regulatory networkds from gene expression data by path consistency algorithm based on conditional mutual information

Pethel, S.D. and Hahs, D.W. (2014). Exact Test of Independence Using Mutual Information
}
\author{
Tong Yin
}
