% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_network.R
\name{functional_enrichment_of_core_circuit}
\alias{functional_enrichment_of_core_circuit}
\title{Functional analysis of TF circuits targets}
\usage{
functional_enrichment_of_core_circuit(TF_network,core_TF_clusters,function_gmt)
}
\arguments{
\item{TF_network}{TF-target result from \code{TF_target_network} function}

\item{core_TF_clusters}{Two-column data.frame define the cluster of each TF}

\item{function_gmt}{A list object containing the vectors of gene list in different functional gene sets}
}
\value{
A data.frame containing function name, overlapping number, enrichment p-value between gene sets and TF targets.
}
\description{
Functional analysis of TF circuits targets
}
