% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_genomic_features.R
\name{hmm_overlapping}
\alias{hmm_overlapping}
\title{Count the number of DRE overlapping with chromatin states}
\usage{
hmm_overlapping(direct_info,CpGs_annotation,chromcol=9)
}
\arguments{
\item{direct_info}{Regulation information of CpG gene pairs detected by MICMIC.
Additional type column should be added to designate the type of regulation.}

\item{CpGs_annotation}{Annotation data frame of each CpG site.}

\item{chromcol}{The column number of chromatin state data in CpGs_annotation data.frame.}
}
\value{
The counting of CpGs located in each chromatin state
}
\description{
Count the number of DRE overlapping with chromatin states
}
