% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICMIC_plotting.R
\name{MICMIC_plotting}
\alias{MICMIC_plotting}
\title{MICMIC_plotting}
\usage{
MICMIC_plotting(gene_name,met_data_matrix,exp_data_matrix,control_id,
distance=NA,ref_gene_bed,ref_CpGs_bed,sample_class,outfiledir=NA)
}
\arguments{
\item{gene_name}{The name of target gene to be plotted}

\item{met_data_matrix}{a SummarizedExperiment object or a numeric matrix containing CpGs methylation data where columns contain samples and rows contain variables(probe site)}

\item{exp_data_matrix}{a SummarizedExperiment object or a numeric matrix containing gene expression data where columns contain samples and rows contain variables(gene site)}

\item{control_id}{a vector containing the ids of control/normal samples}

\item{distance}{Integer specifying the upstream/downstream genome range to be plotted. By default distance will cover all CpGs in analysis result}

\item{ref_gene_bed}{a GRanges object or a data.frame containing reference gene coorinate with five columns named "name", "chr", "start", "end" and "strand". The coordinates of genes in exp_data_matrix are required to be included in this data.frame.}

\item{ref_CpGs_bed}{a GRanges object or a data.frame containing reference CpGS coorinate with four columns names "name", "chr", "start" and "end". The coordinates of CpGs/probes in met_data_matrix are required to be included in this data.frame.}

\item{sample_class}{a data.frame containing the class information for samples}

\item{outfiledir}{a string of file directory to store the result files. If the parameter is not specified, the log file directory will be get by \code{getwd()}.}

\item{network}{plot the direct regulator network or the whole network}
}
\value{
a ggplot object
}
\description{
This function is to map genome coordinates to plotting coordinates, and enlarge target gene promoter and gene body
}
\examples{
data("TCGA_STAD_data")
gene_name<-"MLH1"
\dontrun{
MICMIC_plotting(gene_name=gene_name,met_data_matrix=STAD_met_data_matrix,
exp_data_matrix=STAD_exp_data_matrix,control_id=STAD_control_id,
distance=350000,ref_gene_bed=STAD_ref_gene_bed,
ref_CpGs_bed=STAD_ref_CpGs_bed,sample_class=sample_class)
}
}
\seealso{
the usage of CMI_met_cis_network
}
\author{
Tong Yin
}
