% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMI_GRP_pro.R
\name{PC_para}
\alias{PC_para}
\title{parallel PC network construction based on MI/CMI testing}
\usage{
PC_para(data_matrix,max_L=1,method=c("CMII","CMI"),pre_adj=NULL
,log_file_dir=NA,edgemode=c("pvalue","MI"),pvalue_cut=0.01,core_num=1,permutation_times=100)
}
\arguments{
\item{data_matrix}{a numeric data matrix containing data from observation where columns contain samples(observing) and rows contain variables}

\item{max_L}{The max L of PC. The default value is 1, and that means the network will be infered by CMI testing. If the value is 0, the network will be infered by MI testing.}

\item{method}{choose a  to test interaction between nodes based on conditional mutual information (CMI), or conditional mutual inclusive information.}

\item{pre_adj}{the pre-defined adjacent matrix, representing the hypothetical network. The default value is NULL, and that means all nodes are considered to have association between each other in original hypothesis.}

\item{log_file_dir}{a string of file directory to store the log files. If the parameter is not specified, the log file directory will be get by \code{getwd()}.}

\item{edgemode}{a string value to select the mode in edge decision}

\item{pvalue_cut}{the cutoff of pvalue. The default is 0.01.}

\item{core_num}{the number of CPUs using in the computation.}

\item{permutation_times}{the number of times of permutation to calculate the pvalue}
}
\value{
the adjacency matrix of the network with value of 0 and 1. 1 means that there is an edge between the rowname and colname of the element.
And 0 means there is no edge.
}
\description{
\code{PC_para} is a parallel computation method to infer direct correlation network from data matrix. This method is based on PC-algorithm by conditional mutual information
It will generate an adjacent matrix of the infered network.
}
\examples{
x=rnorm(300,mean=20,sd=6)
y=x+rnorm(300,mean=0,sd=2)
w=y*0.1+rnorm(300,mean=18,sd=1)
v=y*0.15+rnorm(300,mean=17,sd=1)
z=2*w+v+rnorm(300,mean=0,sd=0.1)
a=rnorm(300,mean=20,sd=2)
b=0.9*a+rnorm(300,mean=2,sd=1)
c=b-rnorm(300,mean=0,sd=2)
mydata<-rbind(x,y,w,v,z,a,b,c)
MI_PC_net<-PC_para(mydata,max_L=0,log_file_dir=tempdir())
CMI_PC_net<-PC_para(mydata,max_L=1,method="CMI"log_file_dir=tempdir())
}
\references{
Zhang, X. (2011). Inferring gene regulatory networkds from gene expression data by path consistency algorithm based on conditional mutual information

Zhang, X. (2015). Conditional mutual inclusive information enables accurate quatification of associations in gene regulatory networks.

Kalisch, M. and Buhlmann, P.(2007) Estimating High-Dimensional Directed Acyclic Graphs with the PC-Algorithm.

Pethel, S.D. and Hahs, D.W. (2014). Exact Test of Independence Using Mutual Information
}
\author{
Tong Yin
}
