% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_network.R
\name{plot_core_circuit}
\alias{plot_core_circuit}
\title{Plotting the self-regulation network of the TF circuit}
\usage{
plot_core_circuit(TF_circuit_core,TF_class_data,TF_pathway_data,outpath)
}
\arguments{
\item{TF_circuit_core}{The data.frame of self-regulation relationship from \code{select_TF_circuit} function}

\item{TF_class_data}{Two-column data.frame define the cluster of each TF}

\item{TF_pathway_data}{The result from functional analysis \code{functional_enrichment_of_core_circuit}}

\item{outpath}{The output directory for plotting file.}
}
\value{
A network figure representing the self-regulation relationship between TFs and their functions.
}
\description{
Plotting the self-regulation network of the TF circuit
}
