% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approach_genomic_features.R
\name{plot_histone_mark_score}
\alias{plot_histone_mark_score}
\title{Plot of average histone mark score around CpGs detected by MICMIC}
\usage{
plot_histone_mark_score(direct_info,name,CpGs_score,colors,range,figure_dir)
}
\arguments{
\item{direct_info}{Regulation information of CpG gene pairs detected by MICMIC.
Additional type column should be added to designate the type of regulation.}

\item{name}{The name of histone mark.}

\item{CpGs_score}{The data matrix of histone score around each CpG site.}

\item{colors}{The plotting colors of types of DRE}

\item{range}{The plotting range (bp) around the CpG site.}

\item{figure_dir}{The output directory.}

\item{CpGs_annotation}{Annotation data frame of each CpG site.}
}
\value{
A plotting result would be saved in the output directory to a file named by the histone mark.
}
\description{
Plot of average histone mark score around CpGs detected by MICMIC
}
