% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_network.R
\name{select_TF_circuit}
\alias{select_TF_circuit}
\title{Filtering the TF circuit by odds ratio cutoff and super enhancer annotation}
\usage{
select_TF_circuit(TF_circuit,OR=1.05,super_enhancer=TRUE)
}
\arguments{
\item{TF_circuit}{TF circuit results from \code{selfconnected_TF_circuit} function}

\item{OR}{Odds ratio cutoff}

\item{super_enhancer}{A logical value to decide whether the TF binding CpGs in the circuit
should be all located in super enhancer according to previous studies.}
}
\value{
A data.frame containing the filtered self-regulation relationship within TFs.
}
\description{
Filtering the TF circuit by odds ratio cutoff and super enhancer annotation
}
