% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_network.R
\name{selfconnected_TF_circuit}
\alias{selfconnected_TF_circuit}
\title{Identify the self-connected TF circuit}
\usage{
selfconnected_TF_circuit(TF_binding,CpG_target,CpGs_in_SE)
}
\arguments{
\item{TF_binding}{TF binding results from \code{TF_binding_enrichment} function}

\item{CpG_target}{Two-column data.frame containing CpG-target regulation pairs.}

\item{CpGs_in_SE}{Two-column data.frame annotating CpGs in super enhancer}
}
\value{
A data.frame containing the self-regulation relationship within TFs.
}
\description{
Identify the self-connected TF circuit
}
