% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_network.R
\name{TF_binding_enrichment}
\alias{TF_binding_enrichment}
\title{Calculate the TF binding enrichment on CpG sites}
\usage{
TF_binding_enrichment(motif_data,CpGs_list,control_list)
}
\arguments{
\item{motif_data}{Two-column data.frame containing the TF motif name in the first,
and CpG IDs surrounding the TF binding sites in the second column. CpG IDs should be
separated by ";".}

\item{CpGs_list}{A vector of CpG IDs to analyze.}

\item{control_list}{A vector of randomly selected CpG IDs as background control.}
}
\value{
The odds ratio and confidence interval of each TF on the set the CpG sites.
}
\description{
Calculate the TF binding enrichment on CpG sites
}
