% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_network.R
\name{TF_target_network}
\alias{TF_target_network}
\title{Generate the TF-target network}
\usage{
TF_target_network(TF_binding,CpG_target)
}
\arguments{
\item{TF_binding}{TF binding results from \code{TF_binding_enrichment} function}

\item{CpG_target}{Two-column data.frame containing CpG-target regulation pairs.}
}
\value{
A two-column data.frame containing the TF-target pairs.
}
\description{
Generate the TF-target network
}
